+++ pytsite/http/_request.py
@@ -1,5 +1,6 @@
 import re as _re
 from werkzeug.wrappers import Request as _Request
+from werkzeug.utils import cached_property as _cached_property
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -9,7 +10,7 @@ __license__ = 'MIT'
 class Request(_Request):
     """HTTP request.
     """
-    @property
+    @_cached_property
     def inp(self) -> dict:
         r = {}